+++ lang/cs/global.php
@@ -20,6 +20,7 @@ return [
         'search-hint' => 'Vyhledává se ve sloupcích: :fields',
         'range-from' => ':label od',
         'range-to' => ':label do',
+        'invalid-date' => 'Zadejte datum, např. 2026-10-15, 2026-10, 2026 nebo 3 days ago.',
     ],
     'sort' => [
         'by' => 'Řadit podle',